refact: 添加测试用例
